refactor: side navigation 컴포넌트 분리
